requirements.txt
python-docx>=0.8.11
markdown>=3.4.4
beautifulsoup4>=4.12.0